+++ schedule/permission_unix.go
@@ -11,17 +11,6 @@ import (
 	"github.com/creativeprojects/resticprofile/constants"
 )
 
-// getSchedulePermission returns the permission defined from the configuration,
-// or the best guess considering the current user permission.
-// If the permission can only be guessed, this method will also display a warning
-func getSchedulePermission(permission string) string {
-	permission, safe := detectSchedulePermission(permission)
-	if !safe {
-		clog.Warningf("you have not specified the permission for your schedule (\"system\" or \"user\"): assuming %q", permission)
-	}
-	return permission
-}
-
 // detectSchedulePermission returns the permission defined from the configuration,
 // or the best guess considering the current user permission.
 // safe specifies whether a guess may lead to a too broad or too narrow file access permission.
